refactor: 边框3 tsx 改造
