+++ packages/datav-vue3/components/BorderBox3/index.ts
@@ -1,5 +1,5 @@
 import type { App, Plugin } from 'vue'
-import BorderBox3 from './src/index.vue'
+import BorderBox3 from './src/BorderBox3'
 
 export const BorderBox3Plugin: Plugin = {
   install(app: App) {
