+++ packages/datav-vue3/components/BorderBox3/src/index.less
@@ -1,0 +1,31 @@
+.dv-border-box-3 {
+  position: relative;
+  width: 100%;
+  height: 100%;
+
+  .dv-border-svg-container {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    top: 0px;
+    left: 0px;
+
+    & > polyline {
+      fill: none;
+    }
+  }
+
+  .dv-bb3-line1 {
+    stroke-width: 3;
+  }
+
+  .dv-bb3-line2 {
+    stroke-width: 1;
+  }
+
+  .border-box-content {
+    position: relative;
+    width: 100%;
+    height: 100%;
+  }
+}
