--- packages/datav-vue3/components/BorderBox3/src/index.vue
@@ -1,111 +1,0 @@
-<template>
-  <div ref="borderBox3" class="dv-border-box-3">
-    <svg class="dv-border-svg-container" :width="width" :height="height">
-      <polygon
-        :fill="backgroundColor" :points="`
-        23, 23 ${width - 24}, 23 ${width - 24}, ${height - 24} 23, ${height - 24}
-      `"
-      />
-
-      <polyline
-        class="dv-bb3-line1"
-        :stroke="state.mergedColor[0]"
-        :points="`4, 4 ${width - 22} ,4 ${width - 22}, ${height - 22} 4, ${height - 22} 4, 4`"
-      />
-      <polyline
-        class="dv-bb3-line2"
-        :stroke="state.mergedColor[1]"
-        :points="`10, 10 ${width - 16}, 10 ${width - 16}, ${height - 16} 10, ${height - 16} 10, 10`"
-      />
-      <polyline
-        class="dv-bb3-line2"
-        :stroke="state.mergedColor[1]"
-        :points="`16, 16 ${width - 10}, 16 ${width - 10}, ${height - 10} 16, ${height - 10} 16, 16`"
-      />
-      <polyline
-        class="dv-bb3-line2"
-        :stroke="state.mergedColor[1]"
-        :points="`22, 22 ${width - 4}, 22 ${width - 4}, ${height - 4} 22, ${height - 4} 22, 22`"
-      />
-    </svg>
-
-    <div class="border-box-content">
-      <slot />
-    </div>
-  </div>
-</template>
-
-<script lang="ts" setup>
-import autoResize from 'packages/datav-vue3/utils/autoResize'
-import { deepClone, deepMerge } from 'packages/datav-vue3/utils'
-
-const props = defineProps({
-  color: {
-    type: Array,
-    default: () => ([]),
-  },
-  backgroundColor: {
-    type: String,
-    default: 'transparent',
-  },
-})
-const borderBox3 = ref<HTMLElement | null>(null)
-
-const state = reactive({
-  defaultColor: ['#2862b7', '#2862b7'],
-
-  mergedColor: [],
-})
-
-const { width, height, initWH } = autoResize(borderBox3)
-
-defineExpose({
-  initWH,
-})
-
-watch(() => props.color, () => {
-  mergeColor()
-})
-
-onMounted(() => {
-  mergeColor()
-})
-
-function mergeColor() {
-  state.mergedColor = deepMerge(deepClone(state.defaultColor, true), props.color || [])
-}
-</script>
-
-<style lang="less">
-.dv-border-box-3 {
-  position: relative;
-  width: 100%;
-  height: 100%;
-
-  .dv-border-svg-container {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    top: 0px;
-    left: 0px;
-
-    & > polyline {
-      fill: none;
-    }
-  }
-
-  .dv-bb3-line1 {
-    stroke-width: 3;
-  }
-
-  .dv-bb3-line2 {
-    stroke-width: 1;
-  }
-
-  .border-box-content {
-    position: relative;
-    width: 100%;
-    height: 100%;
-  }
-}
-</style>
