+++ packages/docs/docs/Border/BorderBox3/demo.vue
@@ -1,9 +1,13 @@
 <template>
   <div demo-bg>
-    <dv-border-box-3>
+    <dv-border-box3>
       <div dv-bg>
         dv-border-box3
       </div>
-    </dv-border-box-3>
+    </dv-border-box3>
   </div>
 </template>
+
+<script setup lang="ts">
+import { BorderBox3 as DvBorderBox3 } from '@kjgl77/datav-vue3'
+</script>
